+++ packages/h3/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/h3
 
+## 0.1.88
+
+### Patch Changes
+
+- @vercel/node@5.8.3
+
 ## 0.1.87
 
 ### Patch Changes